+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project shall be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 described with adjustments and template in [TSWiki](https://tswiki.corp.saab.se/CHANGELOG.md_Windows_%26_Apps).
 
+## 4.5
+
+### Added
+- Support for uniform normal, single normal shared by geometry
+
 ## 4.4
 
 ### Added